requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
scikit-learn
scikit-image
torch
torchvision
transformers
google-generativeai
python-dotenv
